requirements.txt
# M1: FastAPI + Gradio UI + unified LLM backends (DeepSeek / Ollama)
fastapi>=0.115.0,<1
itsdangerous>=2.0.0,<3
uvicorn[standard]>=0.30.0,<1
gradio>=4.44.0,<7
jinja2>=3.1.0,<4
python-dotenv>=1.0.0
httpx>=0.27.0,<0.29
openai>=1.40.0,<2
redis>=5.0.0,<6
fakeredis>=2.23.0,<3
